Moved ConstraintsLibrary to detail/ (as it is ModelBasedStateSpace dependent)
