+++ moveit_planners/ompl/ompl_interface/CMakeLists.txt
@@ -3,7 +3,6 @@ set(MOVEIT_LIB_NAME moveit_ompl_interface)
 add_library(${MOVEIT_LIB_NAME}
   src/ompl_interface.cpp
   src/planning_context_manager.cpp
-  src/constraints_library.cpp
   src/model_based_planning_context.cpp
   src/parameterization/model_based_state_space.cpp
   src/parameterization/model_based_state_space_factory.cpp
@@ -11,6 +10,7 @@ add_library(${MOVEIT_LIB_NAME}
   src/parameterization/joint_space/joint_model_state_space_factory.cpp
   src/parameterization/work_space/pose_model_state_space.cpp
   src/parameterization/work_space/pose_model_state_space_factory.cpp
+  src/detail/constraints_library.cpp
   src/detail/threadsafe_state_storage.cpp
   src/detail/state_validity_checker.cpp
   src/detail/projection_evaluators.cpp
