+++ moveit_planners/ompl/ompl_interface/src/detail/constraints_library.cpp
@@ -34,7 +34,7 @@
 
 /* Author: Ioan Sucan */
 
-#include <moveit/ompl_interface/constraints_library.h>
+#include <moveit/ompl_interface/detail/constraints_library.h>
 #include <moveit/ompl_interface/detail/constrained_sampler.h>
 #include <moveit/profiler/profiler.h>
 #include <ompl/tools/config/SelfConfig.h>
